+++ docs/CHANGE_LOG.md
@@ -1,6 +1,7 @@
 # Calendar.js - Change Log:
 
 ## Version 2.1.0:
+- Added "year" support for Holidays, which states if a holiday should only occur within one year (defaults to null).
 - 
 
 <br>